python_api/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
asyncpg==0.28.0
click==8.1.7
exceptiongroup==1.1.3
fastapi==0.103.1
h11==0.14.0
idna==3.4
pydantic==2.3.0
pydantic-core==2.6.3
python-dotenv==1.0.0
sniffio==1.3.0
starlette==0.27.0
typing-extensions==4.8.0
uvicorn==0.23.2
